[CBR-150] WIP: Implement rollback
I think this is done, but I intentionally didn't introduce a fix for an
existing bug (filtering out dependent change) (Actually, did write the fix but
commented it out.) I'd like ot have a test for this though and currently we
don't have one (the generator _can_ generate them  but the tests are too slow
to really try many of them). TBC.

Update: wrote a hand-written test and it fails but the fix doesn't fix it. Out
of time now, will need to get back to this.